debugging pspnet

+++ Models/pspnet.py
@@ -115,8 +115,8 @@ if __name__ == '__main__':
     rgb_im
     print(type(rgb_im))
     print(rgb_im.size())
-    net.double()
+    net.float()
     net.cuda()
     rgb_im.cuda()
-    net(rgb_im.double())
+    net(rgb_im.float())
